Move `ClientEntityMap` and `ClientMapping` to `client_entity_map`
It's quite huge.
